rename functions -> util
It's a more accurate name.
