+++ compressor/util.py
